requirements.txt
numpy==1.16.1
scipy==1.2.1
matplotlib==3.0.3
opencv-python==4.0.1.24
Pillow>=6.0.0
requests==2.21.0
ipython
jupyter==1.0.0
